Test thread pool

+++ lib/core/unittest/CJsonOutputStreamWrapperTest.cc
@@ -53,13 +53,13 @@ void task(ml::core::CJsonOutputStreamWrapper& wrapper, int id, int documents) {
 void CJsonOutputStreamWrapperTest::testConcurrentWrites() {
     std::ostringstream stringStream;
 
-    static const size_t WRITERS(1500);
+    static const int WRITERS(1500);
     static const size_t DOCUMENTS_PER_WRITER(10);
     {
         ml::core::CJsonOutputStreamWrapper wrapper(stringStream);
 
         ml::core::CStaticThreadPool tp(100);
-        for (size_t i = 0; i < WRITERS; ++i) {
+        for (int i = 0; i < WRITERS; ++i) {
             tp.schedule([&wrapper, i] { task(wrapper, i, DOCUMENTS_PER_WRITER); });
         }
     }
